requirements.txt
flask
paho-mqtt
pymysql
requests
psutil
wmi
pywin32
